GH Actions: update for branch-cut of PHP 8.4
The branch for the new PHP (8.4) version has been cut, which means that PHP "nightly" is now PHP 8.5, so we need to add explicit builds to continue testing against PHP 8.4.

Builds against PHP 8.4 should also no longer be allowed to fail as this tool is used in CI, so, even though PHP 8.4 has not been released yet, `8.4` is **not** added to the `continue-on-error` setting.

Also see 180 for reasoning about why the explicit PHP 8.4 build is added "early".