PHPUnit 11.3.1 | AssertIsList: sync error message with upstream
PHPUnit 11.3.1 improved the type information in select assertion failure messages.

For the polyfills, this only affects the `AssertIsList` polyfill.

Now the choice was between the following:
* Just update the test(s) to have a different message expectation for PHPUnit 11.3.1+.
* Update the polyfills to mirror the improvement made in PHPUnit upstream.

I've chosen to implement the latter as it makes the message significantly more informative.

This means that on PHPUnit 6.x to 9.x and 11.3.1+, the new, improved message is shown. While on PHPUnit 10.0 - 11.3.0, the "old" message is shown as, in that case, the PHPUnit native `assertIsList()` method is used without the message improvements.

Includes:
* Renaming the `assertIsListFailureDescription()` method parameter from `$other` to `$value` to be more descriptive.
* Updating the tests, including adding an extra test with an anonymous class as that has separate handling for the error message in the polyfill.

Refs:
* https://github.com/sebastianbergmann/phpunit/commit/5e8b79e52717cf330ce1023ec4b255541812ba70
* https://github.com/sebastianbergmann/phpunit/commit/42cfc7d5982eac708fab2a566021959b31a49aa5