README: fix broken links
Previously, the PHPUnit documentation would have a "main" branch/version, which contained the latest docs.
This branch has been removed, which means all the links to the PHPUnit docs in the README are now broken.

This commit fixes this up to point to the `11.5` branch, which is the latest PHPUnit minor.

Note: it will be a pain in the neck to have to continuously update these links whenever a new PHPUnit minor is released, so it would be nice to be able to automate this at some point.

Ref: https://github.com/sebastianbergmann/phpunit-documentation-english/issues/385