Add pull request template
Mostly as I have a tendency to forget to change the target branch to the oldest supported version for a PR when I haven't worked on this repo for a little while ... oops...

Let's hope this reminder will help prevent such snafus in the future.

Includes a tweak to the Remark config as the file name of the template needs to comply with the requirements by GitHub for such templates and this conflicts with the file name expectations set by Remark (by default).

Ref: https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/creating-a-pull-request-template-for-your-repository