Modernizer: add scalar + return type declarations (wherever possible)
In line with the new minimum of PHPUnit 7.x, the assertions in the PHPUnit Polyfills will now have both parameter as well as return type declarations (wherever possible considering the minimum supported PHP version of PHP 7.1).

Additional type declarations will be added in future majors if/when the minimum supported PHP version allows for it.

Includes removing polyfilled inline type checks for assertions which were introduced in PHPUnit with declared types, but for which the polyfills couldn't type the parameters prior to this.

Includes minor adjustments to exception expectations in the tests to allow for the PHP native `TypeError`s - in contrast to the emulated ones.

Notes:
* `AssertObjectProperty`: the `string` type for the `$propertyName` parameter for the `assertObject[Not]HasProperty()` has not been applied as it would invalidate the type check (as we can't enforce `strict_types` for tests using the assertion).

Refs:
* https://github.com/sebastianbergmann/phpunit/commit/852e5405df1f3f8b46ae54963472b6226f8559e5
* https://github.com/sebastianbergmann/phpunit/commit/a7ab2b9b53c88a9020babcd547d6d71279fba087