Dependabot: remove Composer ecosystem
Composer checks have started to run into trouble with the minimum PHP version vs the YoastCS dependency (which has a higher minimum PHP version).

As Dependabot has never made a useful contribution for the Composer dependencies anyway AND there are only a few, we may as well drop it to get rid of the incessant error notifications.

Ref: https://github.com/Yoast/PHPUnit-Polyfills/network/updates/958266171