GH Actions/test: fix rate limiting issue
If a tool version is specified, workflows can run into rate limiting issues with the GitHub API. To get round this, the `GITHUB_TOKEN` needs to be provided.

The `tools` key is also used in the CS workflows, but has, so far, not yielded any problematic builds for those. If needs be, the same tweak can be made to those job steps if those jobs would start to hit rate limits in the future.

Ref: https://github.com/shivammathur/setup-php#wrench-tools-support