GH Actions: prevent more rate limiting issues
Follow up on 244.

This adds the `GITHUB_TOKEN` to get round the rate limiting where necessary.

It also adds `fail-fast` to get `setup-php` to exit with a non-zero exit code if something required in `tools` could not be installed (or other customizations failed).
This prevents running unnecessary interim steps if a later step would never be able to succeed anyway.