GH Actions: PHP nightly is now PHP 8.6
PHP 8.5 has been branched off, which means that PHP "nightly" is now PHP 8.6/9.0 and to continue testing against PHP 8.5, that version needs to be explicitly added to the matrix.

Builds against PHP 8.5 should also no longer be allowed to fail as this tool is used in CI, so, even though PHP 8.5 has not been released yet, `8.5` is **not** added to the `continue-on-error` setting.

Also see 180 for reasoning about why the explicit PHP 8.5 build is added "early".