PHPCS: selectively ignore some flagged PHP function deprecations
The PHPCS native `Generic.PHP.DeprecatedFunctions` sniff will flag calls to deprecated functions.

As of PHP 8.5, PHP has deprecated a number of functions related to previously done resource to object migrations.

For the purposes of this package, we are testing the `assertClosedResource()` polyfill. These tests use a number of deprecated functions, but these tests will only ever be run on PHP versions where the return value from the involved extension was still a resource.

In other words, the PHP 8.5 deprecations will never be hit and the PHPCS sniff flagging them can be ignored.